RavenDB-5244 V4.0 client long running task - Split DocumentSession to partial classes - Complete unimplemented overloads of load and query - Add summary comments
