+++ src/Raven.Client/Documents/DocumentSession/DocumentSession.Includes.cs
@@ -1,0 +1,47 @@
+//-----------------------------------------------------------------------
+// <copyright file="DocumentSession.cs" company="Hibernating Rhinos LTD">
+//     Copyright (c) Hibernating Rhinos LTD. All rights reserved.
+// </copyright>
+//-----------------------------------------------------------------------
+
+using System;
+using System.Linq.Expressions;
+
+namespace Raven.Client.Documents
+{
+    /// <summary>
+    /// Implements Unit of Work for accessing the RavenDB server
+    /// </summary>
+    public partial class DocumentSession : InMemoryDocumentSessionOperations, IDocumentQueryGenerator, ISyncAdvancedSessionOperation, IDocumentSessionImpl
+    {
+        /// <summary>
+        /// Begin a load while including the specified path
+        /// </summary>
+        /// <param name="path">The path.</param>
+        /// <returns></returns>
+        public ILoaderWithInclude<T> Include<T>(Expression<Func<T, object>> path)
+        {
+            return new MultiLoaderWithInclude<T>(this).Include(path);
+        }
+
+        /// <summary>
+        /// Begin a load while including the specified path
+        /// </summary>
+        /// <param name="path">The path.</param>
+        /// <returns></returns>
+        public ILoaderWithInclude<T> Include<T, TInclude>(Expression<Func<T, object>> path)
+        {
+            return new MultiLoaderWithInclude<T>(this).Include<TInclude>(path);
+        }
+
+        /// <summary>
+        /// Begin a load while including the specified path
+        /// </summary>
+        /// <param name="path">The path.</param>
+        /// <returns></returns>
+        public ILoaderWithInclude<object> Include(string path)
+        {
+            return new MultiLoaderWithInclude<object>(this).Include(path);
+        }
+    }
+}
