+++ src/Raven.Client/Documents/DocumentSession/DocumentSession.cs
@@ -1,0 +1,237 @@
+//-----------------------------------------------------------------------
+// <copyright file="DocumentSession.cs" company="Hibernating Rhinos LTD">
+//     Copyright (c) Hibernating Rhinos LTD. All rights reserved.
+// </copyright>
+//-----------------------------------------------------------------------
+
+using Raven.Client.Connection;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading.Tasks;
+using Raven.Abstractions.Data;
+using Raven.Client.Data;
+using Raven.Client.Data.Queries;
+using Raven.Client.Document;
+using Raven.Client.Documents.Commands;
+using Raven.Client.Documents.SessionOperations;
+using Raven.Client.Http;
+using Raven.Client.Indexes;
+using Raven.Json.Linq;
+using Sparrow.Json;
+
+namespace Raven.Client.Documents
+{
+    /// <summary>
+    /// Implements Unit of Work for accessing the RavenDB server
+    /// </summary>
+    public partial class DocumentSession : InMemoryDocumentSessionOperations, IDocumentQueryGenerator, ISyncAdvancedSessionOperation, IDocumentSessionImpl
+    {
+        /// <summary>
+        /// Gets the database commands.
+        /// </summary>
+        /// <value>The database commands.</value>
+        public IDatabaseCommands DatabaseCommands { get; private set; }
+
+        /// <summary>
+        /// Get the accessor for advanced operations
+        /// </summary>
+        /// <remarks>
+        /// Those operations are rarely needed, and have been moved to a separate 
+        /// property to avoid cluttering the API
+        /// </remarks>
+        public ISyncAdvancedSessionOperation Advanced => this;
+
+        /// <summary>
+        /// Access the lazy operations
+        /// </summary>
+        public ILazySessionOperations Lazily => this;
+
+        /// <summary>
+        /// Access the eager operations
+        /// </summary>
+        public IEagerSessionOperations Eagerly => this;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="DocumentSession"/> class.
+        /// </summary>
+        public DocumentSession(string dbName, DocumentStore documentStore, Guid id, IDatabaseCommands databaseCommands, RequestExecuter requestExecuter)
+            : base(dbName, documentStore, requestExecuter, id)
+        {
+            DatabaseCommands = databaseCommands;
+        }
+        
+        #region Stream
+
+        public IEnumerator<StreamResult<T>> Stream<T>(IQueryable<T> query)
+        {
+            throw new NotImplementedException();
+        }
+
+        public IEnumerator<StreamResult<T>> Stream<T>(IQueryable<T> query, out QueryHeaderInformation queryHeaderInformation)
+        {
+            throw new NotImplementedException();
+        }
+
+        public IEnumerator<StreamResult<T>> Stream<T>(IDocumentQuery<T> query)
+        {
+            throw new NotImplementedException();
+        }
+
+        IEnumerator<StreamResult<T>> ISyncAdvancedSessionOperation.Stream<T>(IDocumentQuery<T> query, out QueryHeaderInformation queryHeaderInformation)
+        {
+            throw new NotImplementedException();
+        }
+
+        public IEnumerator<StreamResult<T>> Stream<T>(IDocumentQuery<T> query, out QueryHeaderInformation queryHeaderInformation)
+        {
+            throw new NotImplementedException();
+        }
+
+        public IEnumerator<StreamResult<T>> Stream<T>(long? fromEtag, int start = 0, int pageSize = Int32.MaxValue,
+            RavenPagingInformation pagingInformation = null, string transformer = null, Dictionary<string, RavenJToken> transformerParameters = null)
+        {
+            throw new NotImplementedException();
+        }
+
+        public IEnumerator<StreamResult<T>> Stream<T>(string startsWith, string matches = null, int start = 0, int pageSize = Int32.MaxValue,
+            RavenPagingInformation pagingInformation = null, string skipAfter = null, string transformer = null,
+            Dictionary<string, RavenJToken> transformerParameters = null)
+        {
+            throw new NotImplementedException();
+        }
+
+        #endregion Stream
+
+        #region DeleteByIndex
+
+        Operation ISyncAdvancedSessionOperation.DeleteByIndex<T>(string indexName, Expression<Func<T, bool>> expression)
+        {
+            return DeleteByIndex(indexName, expression);
+        }
+
+        public Operation DeleteByIndex<T, TIndexCreator>(Expression<Func<T, bool>> expression) where TIndexCreator : AbstractIndexCreationTask, new()
+        {
+            throw new NotImplementedException();
+        }
+
+        public Operation DeleteByIndex<T>(string indexName, Expression<Func<T, bool>> expression)
+        {
+            throw new NotImplementedException();
+        }
+
+        #endregion
+
+        /// <summary>
+        /// Saves all the changes to the Raven server.
+        /// </summary>
+        public void SaveChanges()
+        {
+            var saveChangesOeration = new BatchOperation(this);
+
+            var command = saveChangesOeration.CreateRequest();
+            if (command != null)
+            {
+                RequestExecuter.Execute(command, Context);
+                saveChangesOeration.SetResult(command.Result);
+            }
+        }
+
+        /// <summary>
+        /// Refreshes the specified entity from Raven server.
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="entity">The entity.</param>
+        public void Refresh<T>(T entity)
+        {
+            DocumentInfo documentInfo;
+            if (DocumentsByEntity.TryGetValue(entity, out documentInfo) == false)
+                throw new InvalidOperationException("Cannot refresh a transient instance");
+            IncrementRequestCount();
+
+            //TODO - Change when we have new DatabaseCommands.Get
+            var document = TempDatabaseCommandGet(documentInfo);
+
+            RefreshInternal(entity, document, documentInfo);
+        }
+        
+        private BlittableJsonReaderObject TempDatabaseCommandGet(DocumentInfo documentInfo)
+        {
+            var command = new GetDocumentCommand
+            {
+                Ids = new[] {documentInfo.Id}
+            };
+            RequestExecuter.Execute(command, Context);
+            var document = (BlittableJsonReaderObject) command.Result.Results[0];
+            if (document == null)
+                throw new InvalidOperationException("Document '" + documentInfo.Id +
+                                                    "' no longer exists and was probably deleted");
+
+            object value;
+            document.TryGetMember(Constants.Metadata.Key, out value);
+            documentInfo.Metadata = value as BlittableJsonReaderObject;
+
+            object etag;
+            document.TryGetMember(Constants.Metadata.Etag, out etag);
+            documentInfo.ETag = etag as long?;
+
+            documentInfo.Document = document;
+            return document;
+        }
+
+        /// <summary>
+        /// Gets the document URL for the specified entity.
+        /// </summary>
+        /// <param name="entity">The entity.</param>
+        /// <returns></returns>
+        public string GetDocumentUrl(object entity)
+        {
+            throw new NotImplementedException();
+        }
+         
+        public FacetedQueryResult[] MultiFacetedSearch(params FacetQuery[] queries)
+        {
+            throw new NotImplementedException();
+        }
+
+        /// <summary>
+        /// Gets the document info for the specified document id.
+        /// </summary>
+        /// <param name="documentId">Document Id</param>
+        /// <returns></returns>
+        protected override DocumentInfo GetDocumentInfo(string documentId)
+        {
+            var documentInfo = new DocumentInfo {Id = documentId};
+            DocumentsById.Add(documentId, documentInfo);
+            //TODO - Change when we have new DatabaseCommands.Get
+            TempDatabaseCommandGet(documentInfo);
+            return documentInfo;
+        }
+
+        /// <summary>
+        /// Generates the document key.
+        /// </summary>
+        /// <param name="entity">The entity.</param>
+        /// <returns></returns>
+        protected override string GenerateKey(object entity)
+        {
+            return Conventions.GenerateDocumentKey(DatabaseName, DatabaseCommands, entity);
+        }
+
+        /// <summary>
+        /// Not supported on sync session.
+        /// </summary>
+        /// <param name="entity">The entity.</param>
+        /// <returns></returns>
+        protected override Task<string> GenerateKeyAsync(object entity)
+        {
+            throw new NotSupportedException("Cannot use async operation in sync session");
+        }
+
+        public ResponseTimeInformation ExecuteAllPendingLazyOperations()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
